+++ InputDate/index.jsx
@@ -1,10 +1,69 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import {
+  fontFamily,
+  fontSizeMini,
+} from '../style/font';
+import {
+  white,
+  curiousBlue,
+  outerSpace,
+} from '../style/color';
 import DayPicker from 'react-day-picker';
-import './style.css';
+import NavBar from './InputDateNavBar';
+import Weekday from './InputDateWeekday';
+import Caption from './InputDateCaption';
 import Text from '../Text';
+import moment from 'moment';
 
 /* eslint-disable react/prop-types */
+const datepickerStyles = {
+  display: 'flex',
+  flexWrap: 'wrap',
+  justifyContent: 'center',
+  position: 'relative',
+  userSelect: 'none',
+  fontFamily: fontFamily,
+  minWidth: '336px',
+};
+
+const modifierStyles = {
+  cell: {
+    display: 'inline-block',
+    padding: '.6rem 0rem',
+    width: '3rem',
+    textAlign: 'center',
+    cursor: 'pointer',
+    verticalAlign: 'middle',
+    overflow: 'hidden',
+    color: 'black',
+    fontSize: fontSizeMini,
+    border: '0.5px solid #eff1f1',
+    outline: '0.5px solid #eff1f1',
+    boxSizing: 'border-box',
+  },
+  today: {
+    color: curiousBlue,
+  },
+  outside: {
+    color: '#dce0e0',
+    cursor: 'default',
+  },
+  disabled: {
+    color: '#dce0e0',
+    cursor: 'default',
+    backgroundColor: '#f8f8f8',
+  },
+  selected: {
+    color: white,
+    backgroundColor: outerSpace,
+    width: 'calc(100% / 7)',
+  }
+};
+
+const modifiers = {
+  cell: { daysOfWeek: [0, 1, 2, 3, 4, 5, 6] },
+};
 
 const renderError = ({ error, touched }) => (
   error && touched ? (
@@ -12,6 +71,7 @@ const renderError = ({ error, touched }) => (
       style={{
         marginTop: '1rem',
         textAlign: 'center',
+        width: '100%',
       }}
     >
       <Text color={'torchRed'}>{ error }</Text>
@@ -19,7 +79,7 @@ const renderError = ({ error, touched }) => (
   ) : null
 );
 
-const disabledDays = ({ disableBefore, submitting }) => {
+const disabledDays = ({ disableBefore, submitting }, initialMonth) => {
   if (submitting) {
     return () => true;
   } else if (disableBefore) {
@@ -27,10 +87,12 @@ const disabledDays = ({ disableBefore, submitting }) => {
       before: new Date(disableBefore.year, disableBefore.month, disableBefore.day),
     };
   }
+  return {
+    before: initialMonth,
+  }
 };
 
 /* eslint-enable react/prop-types */
-
 
 const InputDate = ({
   disableBefore,
@@ -45,12 +107,30 @@ const InputDate = ({
     submitting,
   },
   firstDayOfWeek,
+  firstMonthToDisplay,
+  onNavigationClick,
+  weekdayLength,
+  date,
+  renderDay,
+  onDayClick,
+  selectedDays,
+  initialMonth,
 }) =>
-  <div>
+  <div style={datepickerStyles}>
     <DayPicker
+      navbarElement={<NavBar firstMonthToDisplay={firstMonthToDisplay} onNavigationClick={onNavigationClick}/>}
+      weekdayElement={<Weekday weekdayLength={weekdayLength} />}
+      captionElement={<Caption date={date}/>}
       className={submitting ? 'disabled' : undefined}
-      disabledDays={disabledDays({ disableBefore, submitting })}
-      initialMonth={new Date(initialMonthYear.year, initialMonthYear.month)}
+      disabledDays={disabledDays({ disableBefore, submitting }, initialMonth)}
+      // First month the user can navigate to, default to current month:
+      fromMonth={firstMonthToDisplay}
+      // Month showed when open the calendar:
+      initialMonth={initialMonthYear? new Date(initialMonthYear.year, initialMonthYear.month) : initialMonth}
+      renderDay={renderDay}
+      modifiers={modifiers}
+      modifiersStyles={modifierStyles}
+      showOutsideDays
       onDayClick={(day, { disabled }) => {
         if (!disabled) {
           onChange({
@@ -58,9 +138,10 @@ const InputDate = ({
             month: day.getMonth(),
             year: day.getFullYear(),
           });
+          onDayClick(day);
         }
       }}
-      selectedDays={!value ? null : new Date(value.year, value.month, value.day)}
+      selectedDays={selectedDays ? selectedDays : (!value ? null : new Date(value.year, value.month, value.day))}
       firstDayOfWeek={firstDayOfWeek}
     />
     {renderError({
@@ -84,7 +165,7 @@ InputDate.propTypes = {
   initialMonthYear: PropTypes.shape({
     month: PropTypes.number,
     year: PropTypes.number,
-  }).isRequired,
+  }),
   disableBefore: PropTypes.shape({
     month: PropTypes.number,
     year: PropTypes.number,
@@ -95,11 +176,26 @@ InputDate.propTypes = {
     submitting: PropTypes.bool,
   }),
   firstDayOfWeek: PropTypes.oneOf([0, 1, 2, 3, 4, 5, 6]),
+  firstMonthToDisplay: PropTypes.instanceOf(Date),
+  initialMonth: PropTypes.instanceOf(Date),
+  onNavigationClick: PropTypes.func,
+  weekdayLength: PropTypes.oneOf(['short', 'medium', 'long']),
+  date: PropTypes.instanceOf(Date),
+  renderDay: PropTypes.func,
+  onDayClick: PropTypes.func,
 };
 
 InputDate.defaultProps = {
+  input: {
+    value: '',
+    onChange: () => {},
+  },
   meta: {},
   firstDayOfWeek: 0,
+  firstMonthToDisplay: moment().toDate(),
+  onNavigationClick: () => {},
+  weekdayLength: 'short',
+  onDayClick: () => {},
 };
 
 export default InputDate;