+++ Popover/index.jsx
@@ -8,49 +8,59 @@ import {
   modal,
 } from '../style/zIndex';
 
+
 const Popover = ({
   children,
   left,
   top,
   bottom,
   right,
+  width,
   onOverlayClick,
   transparentOverlay,
-}) =>
-  <span>
-    <div
-      style={calculateStyles({
-        default: {
-          position: 'absolute',
-          left,
-          top,
-          bottom,
-          right,
-          zIndex: modal,
-        },
-        center: {
-          top: '50%',
-          left: '50%',
-          transform: 'translate(-50%, -50%)',
-        },
-      }, {
-        center: !(left || top || bottom || right),
-      })}
-    >
-      {children}
+}) => {
+  const popoverWrapperStyle = {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    display: 'flex',
+    flexDirection: 'column',
+    height: '100vh',
+    width: `${width || '100vw'}`,
+    justifyContent: 'center',
+    alignItems: 'center',
+  };
+  return (
+    <div style={popoverWrapperStyle}>
+      <div
+        style={calculateStyles({
+          default: {
+            position: 'absolute',
+            left,
+            top,
+            bottom,
+            right,
+            zIndex: modal,
+          },
+        })}
+      >
+        {children}
+      </div>
+      <Overlay
+        onClick={onOverlayClick}
+        transparent={transparentOverlay}
+      />
     </div>
-    <Overlay
-      onClick={onOverlayClick}
-      transparent={transparentOverlay}
-    />
-  </span>;
+  );
+};
 
 Popover.propTypes = {
-  children: PropTypes.node,
+  children: PropTypes.node.isRequired,
   left: PropTypes.string,
   top: PropTypes.string,
   bottom: PropTypes.string,
   right: PropTypes.string,
+  width: PropTypes.string,
   onOverlayClick: PropTypes.func,
   transparentOverlay: PropTypes.bool,
 };