+++ Video/story.jsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { checkA11y } from 'storybook-addon-a11y';
 import Video from './index';
 
-storiesOf('Video')
+storiesOf('Video', module)
   .addDecorator(checkA11y)
   .add('Default', () => (
     <Video